Update for current JWT lib
Update to account for changes in the current Firebase JWT library.
You're required to pass an array of allowed decryption algorithms when
calling JWT::decode().

In addition, update the expiration date on the example token to be 2030
so the example will continue to work for a long time. And store the
decrypted JWT in the code so that it's easier to debug. Also, this
provides an example of how to encrypt a JWT.